requirements.txt
Flask>=3.0.0
Authlib>=1.2.0
python-dotenv>=1.0.0
requests